+++ fake_craft_cli/__init__.py
@@ -1,4 +1,4 @@
-# Copyright 2020-2021 Canonical Ltd.
+# Copyright 2021 Canonical Ltd.
 #
 # Licensed under the Apache License, Version 2.0 (the "License");
 # you may not use this file except in compliance with the License.
@@ -14,12 +14,7 @@
 #
 # For further info, check https://github.com/canonical/charmcraft
 
-"""Infrastructure for common base commands functionality."""
+"""Present package API."""
 
-from craft_cli import CraftError
-
-
-class CommandError(CraftError):
-    """Base exception for all error commands."""
-    # XXX 2021-12-14 Facundo: this is kept to avoid massive renaming in all code, which
-    # we may do sometime in the future
+from .errors import ArgumentParsingError, ProvideHelpException  # noqa
+from .dispatcher import GlobalArgument, Dispatcher, BaseCommand  # noqa
